Add validation for Swift property names
Swift properties which fall in to any of the ARC method families don't work
properly via KVC, so we can't support them and will either leak memory or crash
when trying to use them (currently we crash).

Trying to use properties with these names is a compilation error in
Objective-C, so it's not an issue there.